Rename tasks/process to tasks/upload, reorganize
