+++ multinet/api/tasks/upload/common.py
@@ -1,0 +1,6 @@
+from multinet.api.models import Upload
+from multinet.api.tasks import MultinetCeleryTask
+
+
+class ProcessUploadTask(MultinetCeleryTask):
+    task_model = Upload
