+++ multinet/api/tasks/upload/csv.py
@@ -7,7 +7,8 @@ from celery.utils.log import get_task_logger
 
 from multinet.api.models import Table, Upload
 
-from .utils import ColumnTypeEnum, ProcessUploadTask, processor_dict
+from .common import ProcessUploadTask
+from .utils import ColumnTypeEnum, processor_dict
 
 logger = get_task_logger(__name__)
