+++ multinet/api/views/serializers.py
@@ -3,7 +3,7 @@ from django.contrib.auth.validators import UnicodeUsernameValidator
 from rest_framework import serializers
 
 from multinet.api.models import Network, Table, Upload, Workspace
-from multinet.api.tasks.process.utils import ColumnTypeEnum
+from multinet.api.tasks.upload.utils import ColumnTypeEnum
 
 
 # The default ModelSerializer for User fails if the user already exists
